conditionally render expense table on button click
